fix tests to add authentication to routers

+++ tests/test_tour_routes.py
@@ -1,28 +1,25 @@
 import json
 import os
-from datetime import datetime
 from pathlib import Path
-from unittest.mock import patch
+from urllib.parse import urlencode
+from dateutil.parser import parse
 
 import pytest
 from async_asgi_testclient import TestClient
 from odmantic.engine import AIOEngine
 
-from natours.models.tour_model import Tours
+from natours.models.tour_model import Tour
 
 pytestmark = pytest.mark.asyncio
 
 
-@pytest.fixture
-async def test_client(engine: AIOEngine) -> TestClient:
-    with patch("natours.models.database.engine", engine):
-        from natours.app import app
-
-        async with TestClient(app) as client:
-            yield client
-
-
 # test happy path
+
+def prep_tour(tour:dict):
+    tour["startDates"] = [ parse(d, ignoretz=True).isoformat() for d in tour["startDates"]]
+    tour["id"] = tour["_id"]
+    del tour["_id"]
+    return tour
 
 
 async def test_heart_beat(test_client: TestClient):
@@ -30,35 +27,53 @@ async def test_heart_beat(test_client: TestClient):
     assert response.status_code == 200
     assert response.json() == {"I ❤️ FastAPI": "🙋🏽‍♂️"}
 
+async def test_sign_up(test_client: TestClient):
+    body = {
+            "username":"test",
+            "password":"1234",
+            "confirm_password":"1234",
+            "email": "test@email.com"
+            }
+    response = await test_client.post("/api/v1/users/signup", json=body)
+    response.json()["status"] == "success"
+    response.status_code == 200
 
-async def test_post_tours(test_client: TestClient, engine: AIOEngine):
+async def test_get_token(test_client: TestClient):
+    param = {"username":"test", "password":"1234"}
+
+    headers = {
+    'Content-Type': 'application/x-www-form-urlencoded'
+    }
+    response =  await test_client.post("/api/v1/users/token", data=urlencode(param), headers=headers)
+    assert response.status_code == 200
+
+async def test_get_me(test_client: TestClient, admin_token_header: dict):
+    response = await test_client.get("/api/v1/users/me", headers=admin_token_header)
+    assert response.status_code == 200
+ 
+
+async def test_post_tours(test_client: TestClient, engine: AIOEngine, admin_token_header: dict):
     p = Path(os.path.dirname(__file__)).resolve().parent
-    with open(p / "natours/dev-data/data/tours-simple.json", "r") as fp:
+    with open(p / "natours/dev-data/data/tours.json", "r") as fp:
         data = json.load(fp)
-    tours = [{k: v for k, v in d.items() if k != "id"} for d in data]
-    for tour in tours:
-        tour["startDates"] = [
-            datetime.strptime(d, "%Y-%m-%d,%H:%M").isoformat()
-            for d in tour["startDates"]
-        ]
-        response = await test_client.post("/api/v1/tours/", json=tour)
+    for tour in data:
+        response = await test_client.post("/api/v1/tours/", json=prep_tour(tour), headers=admin_token_header)
         assert response.status_code == 200
-        assert response.json()["status"] == "success"
 
-    fetched_from_db = await engine.find(Tours)
+    fetched_from_db = await engine.find(Tour)
     assert len(fetched_from_db) == 9
 
 
-async def test_get_tours(test_client: TestClient, engine: AIOEngine):
-    response = await test_client.get("/api/v1/tours/")
+async def test_get_tours(test_client: TestClient, engine: AIOEngine, admin_token_header: dict):
+    response = await test_client.get("/api/v1/tours/", headers=admin_token_header)
     assert response.status_code == 200
     assert response.json()["status"] == "success"
     assert response.json()["results"] == 9
-    assert await engine.find(Tours) is not None
+    assert await engine.find(Tour) is not None
 
 
 async def test_get_tour(test_client: TestClient, engine: AIOEngine):
-    tour = await engine.find(Tours)
+    tour = await engine.find(Tour)
     tour_id = tour[0].dict()["id"]
     tour_name = tour[0].dict()["name"]
     tour_difficulty = tour[0].dict()["difficulty"]
@@ -69,22 +84,22 @@ async def test_get_tour(test_client: TestClient, engine: AIOEngine):
     assert data["difficulty"] == tour_difficulty
 
 
-async def test_patch_tour(test_client: TestClient, engine: AIOEngine):
-    tour = await engine.find(Tours)
+async def test_patch_tour(test_client: TestClient, engine: AIOEngine, admin_token_header: dict):
+    tour = await engine.find(Tour)
     tour_id = tour[0].dict()["id"]
     tour_name = tour[0].dict()["name"]
     response = await test_client.patch(
-        f"/api/v1/tours/{tour_id}", json=dict(name="this is the tour new name")
+        f"/api/v1/tours/{tour_id}", json=dict(name="this is the tour new name") , headers=admin_token_header
     )
     assert response.status_code == 200
-    tour = await engine.find(Tours, Tours.id == tour_id)
+    tour = await engine.find(Tour, Tour.id == tour_id)
     assert tour[0].dict()["name"] == "this is the tour new name"
     assert tour[0].dict()["name"] != tour_name
     response = await test_client.patch(
-        f"/api/v1/tours/{tour_id}", json=dict(name=tour_name)
+        f"/api/v1/tours/{tour_id}", json=dict(name=tour_name), headers=admin_token_header
     )
     assert response.status_code == 200
-    tour = await engine.find(Tours, Tours.id == tour_id)
+    tour = await engine.find(Tour, Tour.id == tour_id)
     assert tour[0].dict()["name"] == tour_name
 
 
@@ -93,7 +108,7 @@ async def test_tour_stats(test_client: TestClient, engine: AIOEngine):
     assert response.status_code == 200
     data = response.json()
     assert len(data["tour stats"]) == 3
-    assert sum([stat["numTours"] for stat in data["tour stats"]]) == 9
+    assert sum([stat["numTours"] for stat in data["tour stats"]]) == 7
 
 
 @pytest.mark.parametrize("year, status_code", [(2021, 200), (99999, 404)])
@@ -104,18 +119,17 @@ async def test_monthly_plan(test_client: TestClient, year: int, status_code: int
     print(data)
 
 
-async def test_delete_tour(test_client: TestClient, engine: AIOEngine):
-    tour = await engine.find(Tours)
+async def test_delete_tour(test_client: TestClient, engine: AIOEngine, admin_token_header: dict):
+    tour = await engine.find(Tour)
     tour_id = tour[0].dict()["id"]
-    response = await test_client.delete(f"/api/v1/tours/{tour_id}")
+    response = await test_client.delete(f"/api/v1/tours/{tour_id}", headers=admin_token_header)
     assert response.status_code == 200
-    assert await engine.find(Tours, Tours.id == tour_id) == []
-    tours = await engine.find(Tours)
+    assert await engine.find(Tour, Tour.id == tour_id) == []
+    tours = await engine.find(Tour)
     assert len(tours) == 8
 
 
 # test errors and fails
-
 
 async def test_inexistent_path(test_client: TestClient):
     response = await test_client.get(
@@ -128,7 +142,7 @@ async def test_inexistent_path(test_client: TestClient):
 async def test_get_tour_non_existing_id_but_valid_object_id(
     test_client: TestClient, engine: AIOEngine
 ):
-    tour = await engine.find(Tours)
+    tour = await engine.find(Tour)
     tour_id = tour[0].dict()["id"]
     tour_id_str = f"{tour_id}"
     response = await test_client.get(f"/api/v1/tours/{tour_id_str[:-8]}12345678")
@@ -138,7 +152,7 @@ async def test_get_tour_non_existing_id_but_valid_object_id(
 async def test_get_tour_with_invalid_object_id(
     test_client: TestClient, engine: AIOEngine
 ):
-    tour = await engine.find(Tours)
+    tour = await engine.find(Tour)
     tour_id = tour[0].dict()["id"]
     tour_id_str = f"{tour_id}"
     response = await test_client.get(f"/api/v1/tours/{tour_id_str[:-1]}")
@@ -147,30 +161,30 @@ async def test_get_tour_with_invalid_object_id(
 
 
 async def test_patch_tour_with_invalid_object_id(
-    test_client: TestClient, engine: AIOEngine
+    test_client: TestClient, engine: AIOEngine, admin_token_header: dict
 ):
-    tour = await engine.find(Tours)
+    tour = await engine.find(Tour)
     tour_id = tour[0].dict()["id"]
     tour_id_str = f"{tour_id}"
     response = await test_client.patch(
         f"/api/v1/tours/{tour_id_str[:-4]}4321",
-        json=dict(name="this is the tour new name", duration=444),
+        json=dict(name="this is the tour new name", duration=444) , headers=admin_token_header
     )
     assert response.status_code == 404
     response = await test_client.patch(
         f"/api/v1/tours/{tour_id_str[:-4]}",
-        json=dict(name="this is the tour new name", duration=444),
+        json=dict(name="this is the tour new name", duration=444) , headers=admin_token_header
     )
     assert response.status_code == 422
 
 
 async def test_delete_tour_with_invalid_object_id(
-    test_client: TestClient, engine: AIOEngine
+    test_client: TestClient, engine: AIOEngine, admin_token_header: dict
 ):
-    tour = await engine.find(Tours)
+    tour = await engine.find(Tour)
     tour_id = tour[0].dict()["id"]
     tour_id_str = f"{tour_id}"
-    response = await test_client.delete(f"/api/v1/tours/{tour_id_str[:-4]}4321")
+    response = await test_client.delete(f"/api/v1/tours/{tour_id_str[:-4]}4321", headers=admin_token_header)
     assert response.status_code == 404
-    response = await test_client.delete(f"/api/v1/tours/{tour_id_str[:-4]}")
+    response = await test_client.delete(f"/api/v1/tours/{tour_id_str[:-4]}", headers=admin_token_header)
     assert response.status_code == 422
